+++ packages/ocean-docs/sidebars.ts
@@ -48,6 +48,8 @@ const sidebars: SidebarsConfig = {
         'components/input',
         'components/link',
         'components/list',
+        'components/listaction',
+        'components/listexpandable',
         'components/modal',
         'components/progress',
         'components/radio',